+++ airdcpp/SSLSocket.h
@@ -54,7 +54,7 @@ public:
 	virtual void connect(const Socket::AddressInfo& aIp, const string& aPort);
 	virtual int read(void* aBuffer, int aBufLen);
 	virtual int write(const void* aBuffer, int aLen);
-	virtual std::pair<bool, bool> wait(uint32_t millis, bool checkRead, bool checkWrite);
+	virtual std::pair<bool, bool> wait(uint64_t millis, bool checkRead, bool checkWrite);
 	virtual void shutdown() noexcept;
 	virtual void close() noexcept;
 
@@ -65,8 +65,8 @@ public:
 	virtual ByteVector getKeyprint() const noexcept;
 	virtual bool verifyKeyprint(const string& expKeyp, bool allowUntrusted) noexcept;
 
-	virtual bool waitConnected(uint32_t millis);
-	virtual bool waitAccepted(uint32_t millis);
+	virtual bool waitConnected(uint64_t millis);
+	virtual bool waitAccepted(uint64_t millis);
 
 private:
 
@@ -76,7 +76,7 @@ private:
 	unique_ptr<CryptoManager::SSLVerifyData> verifyData;	// application data used by CryptoManager::verify_callback(...)
 
 	int checkSSL(int ret);
-	bool waitWant(int ret, uint32_t millis);
+	bool waitWant(int ret, uint64_t millis);
 };
 
 } // namespace dcpp
